requirements.txt
#python 3.10.6
asgiref==3.8.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.5
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
